refactor: invert PushItem() dependency to facilitate chunk handler
