+++ handler/datasource/logger.go
@@ -1,0 +1,5 @@
+package datasource
+
+import "github.com/ipfs/go-log/v2"
+
+var logger = log.Logger("singularity/handler/datasource")
